refactor: split bundled webui app
