+++ codexbox/public/index.html
@@ -80,6 +80,9 @@
       </main>
     </div>
 
+    <script src="/static/app-transport.js" defer></script>
+    <script src="/static/app-render.js" defer></script>
+    <script src="/static/app-session.js" defer></script>
     <script src="/static/app.js" defer></script>
   </body>
 </html>
